Made copy URL also copy over potential chunks.
Since most chunk men probably want to keep their list of potential chunks
the same every time, it would be useful to be able to bookmark a list of
not only unlocked chunks, but also potential chunks.
This way, it is much easier to keep track of which chunks are blacklisted,
not unlockable, etc. and which chunks are a potential unlockable chunk
next iteration.

I've tried making as few changes as possible (except for cleaning up the
code that pads a number with zeroes). I haven't really investigated what
the encoded string represents, I've just used the system you already had
in place and expanded it to also include potential chunks.